Clean up the helical approach for the IP
The helical approach now does not need unit conversion and only uses
information from the tracker. This results in not need the 4-momentum of
the particle anymore, since v_z is calculated from the helix parameters.
This also allows include the uncertainty on v_z in the IP Significance,
since it is expressed in helix parameters.